pyproject.toml
[project]
name = "mqtts"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "Wataru Nakata", email = "wataru9871@gmail.com" }
]
dependencies = ["torch~=2.0.1", "lightning~=2.0.3", "torchaudio~=2.0.2", "hydra-core~=1.3.2", "webdataset~=0.2.48", "wandb~=0.15.4", "soundfile~=0.12.1", "phonemizer~=3.2.1", "matplotlib~=3.7.1", "pyopenjtalk[marine]~=0.3.0"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
[tool.hatch.metadata]
allow-direct-references = true
